refactor: move override logic to lib hooks
