+++ packages/core/lib/use-default-render.tsx
@@ -1,0 +1,13 @@
+import { ReactNode, useMemo } from "react";
+
+export const useDefaultRender = () => {
+    return useMemo<
+      React.FunctionComponent<{ children?: ReactNode }>
+    >(() => {
+      const PuckDefault = ({ children }: { children?: ReactNode }) => (
+        <>{children}</>
+      );
+  
+      return PuckDefault;
+    }, []);
+  };
